+++ source/chip/STM32/STM32F7/STM32F7-RCC.cpp
@@ -2,7 +2,7 @@
  * \file
  * \brief Implementation of RCC-related functions for STM32F7
  *
- * \author Copyright (C) 2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2017-2019 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -74,6 +74,11 @@ int configureApbClockDivider(const bool ppre2, const uint8_t ppre)
 	return EINVAL;
 }
 
+void configurePll48ClockSource(const bool pllsaip)
+{
+	RCC->DCKCFGR2 = (RCC->DCKCFGR2 & ~RCC_DCKCFGR2_CK48MSEL) | pllsaip << RCC_DCKCFGR2_CK48MSEL_Pos;
+}
+
 void configurePllClockSource(const bool hse)
 {
 	RCC->PLLCFGR = (RCC->PLLCFGR & ~RCC_PLLCFGR_PLLSRC) | hse << RCC_PLLCFGR_PLLSRC_Pos;
@@ -96,6 +101,16 @@ void disableHse()
 void disablePll()
 {
 	RCC->CR &= ~RCC_CR_PLLON;
+}
+
+void disablePlli2s()
+{
+	RCC->CR &= ~RCC_CR_PLLI2SON;
+}
+
+void disablePllsai()
+{
+	RCC->CR &= ~RCC_CR_PLLSAION;
 }
 
 void enableHse(const bool bypass)
@@ -134,6 +149,79 @@ int enablePll(const uint16_t plln, const uint8_t pllp, const uint8_t pllq)
 
 	RCC->CR |= RCC_CR_PLLON;
 	while ((RCC->CR & RCC_CR_PLLRDY) == 0);	// wait until PLL is stable
+	return 0;
+}
+
+#if defined(CONFIG_CHIP_STM32F72) || defined(CONFIG_CHIP_STM32F73)
+int enablePlli2s(const uint16_t plli2sn, const uint8_t plli2sq, const uint8_t plli2sr)
+#else	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+int enablePlli2s(const uint16_t plli2sn, const uint8_t plli2sp, const uint8_t plli2sq, const uint8_t plli2sr)
+#endif	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+{
+	if (plli2sn < minPlln || plli2sn > maxPlln || plli2sq < minPllq || plli2sq > maxPllq ||
+			plli2sr < minPllr || plli2sr > maxPllr)
+		return EINVAL;
+
+#if defined(CONFIG_CHIP_STM32F72) || defined(CONFIG_CHIP_STM32F73)
+
+	RCC->PLLI2SCFGR = (RCC->PLLI2SCFGR & ~(RCC_PLLI2SCFGR_PLLI2SN | RCC_PLLI2SCFGR_PLLI2SQ | RCC_PLLI2SCFGR_PLLI2SR)) |
+			plli2sn << RCC_PLLI2SCFGR_PLLI2SN_Pos |
+			plli2sq << RCC_PLLI2SCFGR_PLLI2SQ_Pos |
+			plli2sr << RCC_PLLI2SCFGR_PLLI2SR_Pos;
+
+#else	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+
+	if (plli2sp != pllpDiv2 && plli2sp != pllpDiv4 && plli2sp != pllpDiv6 && plli2sp != pllpDiv8)
+		return EINVAL;
+
+	RCC->PLLI2SCFGR = (RCC->PLLI2SCFGR & ~(RCC_PLLI2SCFGR_PLLI2SN | RCC_PLLI2SCFGR_PLLI2SP | RCC_PLLI2SCFGR_PLLI2SQ |
+			RCC_PLLI2SCFGR_PLLI2SR)) |
+			plli2sn << RCC_PLLI2SCFGR_PLLI2SN_Pos |
+			(plli2sp / 2 - 1) << RCC_PLLI2SCFGR_PLLI2SP_Pos |
+			plli2sq << RCC_PLLI2SCFGR_PLLI2SQ_Pos |
+			plli2sr << RCC_PLLI2SCFGR_PLLI2SR_Pos;
+
+#endif	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+
+	RCC->CR |= RCC_CR_PLLI2SON;
+	while ((RCC->CR & RCC_CR_PLLI2SRDY) == 0);	// wait until PLLI2S is stable
+	return 0;
+}
+
+#if defined(CONFIG_CHIP_STM32F72) || defined(CONFIG_CHIP_STM32F73)
+int enablePllsai(const uint16_t pllsain, const uint8_t pllsaip, const uint8_t pllsaiq)
+#else	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+int enablePllsai(const uint16_t pllsain, const uint8_t pllsaip, const uint8_t pllsaiq, const uint8_t pllsair)
+#endif	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+{
+	if (pllsain < minPlln || pllsain > maxPlln ||
+			(pllsaip != pllpDiv2 && pllsaip != pllpDiv4 && pllsaip != pllpDiv6 && pllsaip != pllpDiv8) ||
+			pllsaiq < minPllq || pllsaiq > maxPllq)
+		return EINVAL;
+
+#if defined(CONFIG_CHIP_STM32F72) || defined(CONFIG_CHIP_STM32F73)
+
+	RCC->PLLSAICFGR = (RCC->PLLSAICFGR & ~(RCC_PLLSAICFGR_PLLSAIN | RCC_PLLSAICFGR_PLLSAIP | RCC_PLLSAICFGR_PLLSAIQ)) |
+			pllsain << RCC_PLLSAICFGR_PLLSAIN_Pos |
+			(pllsaip / 2 - 1) << RCC_PLLSAICFGR_PLLSAIP_Pos |
+			pllsaiq << RCC_PLLSAICFGR_PLLSAIQ_Pos;
+
+#else	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+
+	if (pllsair < minPllr || pllsair > maxPllr)
+		return EINVAL;
+
+	RCC->PLLSAICFGR = (RCC->PLLSAICFGR & ~(RCC_PLLSAICFGR_PLLSAIN | RCC_PLLSAICFGR_PLLSAIP | RCC_PLLSAICFGR_PLLSAIQ |
+			RCC_PLLSAICFGR_PLLSAIR)) |
+			pllsain << RCC_PLLSAICFGR_PLLSAIN_Pos |
+			(pllsaip / 2 - 1) << RCC_PLLSAICFGR_PLLSAIP_Pos |
+			pllsaiq << RCC_PLLSAICFGR_PLLSAIQ_Pos |
+			pllsair << RCC_PLLSAICFGR_PLLSAIR_Pos;
+
+#endif	// !defined(CONFIG_CHIP_STM32F72) && !defined(CONFIG_CHIP_STM32F73)
+
+	RCC->CR |= RCC_CR_PLLSAION;
+	while ((RCC->CR & RCC_CR_PLLSAIRDY) == 0);	// wait until PLLSAI is stable
 	return 0;
 }