+++ source/chip/STM32/STM32F7/distortos-sources.cmake
@@ -1,7 +1,7 @@
 #
 # file: distortos-sources.cmake
 #
-# author: Copyright (C) 2018 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+# author: Copyright (C) 2018-2019 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
 # distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -57,12 +57,12 @@ if(distortos_Clocks_00_Standard_configuration_of_clocks)
 	endif(CONFIG_CHIP_STM32F7_RCC_HSE_FREQUENCY)
 
 	distortosSetConfiguration(BOOLEAN
-			distortos_Clocks_04_PLL
+			distortos_Clocks_04_PLLs
 			ON
-			HELP "Enable PLL."
-			OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLL_ENABLE)
+			HELP "Enable PLLs."
+			OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLS_ENABLE)
 
-	if(distortos_Clocks_04_PLL)
+	if(distortos_Clocks_04_PLLs)
 
 		if(distortos_Clocks_03_HSE)
 			set(pllClockSourceHse HSE)
@@ -83,83 +83,248 @@ if(distortos_Clocks_00_Standard_configuration_of_clocks)
 				MAX 63
 				HELP "PLLM value for main and audio PLLs.
 
-				It is used to divide PLL input clock (PLLin) before it is fed to VCO. VCO input frequency (VCOin) must
-				be in the range [1 MHz; 2 MHz], 2 MHz is recommended to limit PLL jitter.
+				It is used to divide PLL input clock (PLLin) before it is fed to VCO of each PLL. VCO input frequency
+				(VCOin) must be in the range [1 MHz; 2 MHz], 2 MHz is recommended to limit PLL jitter.
 
 				VCOin = PLLin / PLLM"
 				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLM)
 
-		distortosSetConfiguration(INTEGER
-				distortos_Clocks_07_PLLN
-				336
-				MIN 50
-				MAX 432
-				HELP "PLLN value for main PLL.
+		distortosSetConfiguration(BOOLEAN
+				distortos_Clocks_07_PLL
+				ON
+				HELP "Enable main PLL."
+				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLL_ENABLE)
 
-				It is used to multiply VCO input frequency (VCOin). Resulting VCO output frequency (VCOout) must be in
-				the range [100 MHz; 432 MHz].
-
-				VCOout = VCOin * PLLN = PLLin / PLLM * PLLN"
-				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLN)
-
-		distortosSetConfiguration(STRING
-				distortos_Clocks_08_PLLP
-				2
-				4
-				6
-				8
-				HELP "PLLP value for main PLL.
-
-				It is used to divide VCO output frequency (VCOout). PLL output frequency (PLLout) must not exceed:
-				- 216 MHz (with over-drive mode) or 180 MHz (otherwise) in voltage scale 1 mode,
-				- 180 MHz (with over-drive mode) or 168 MHz (otherwise) in voltage scale 2 mode,
-				- 144 MHz in voltage scale 3 mode.
-
-				PLLout = VCOout / PLLP = PLLin / PLLM * PLLN / PLLP"
-				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLP
-				OUTPUT_TYPES INTEGER)
-
-		distortosSetConfiguration(INTEGER
-				distortos_Clocks_09_PLLQ
-				7
-				MIN 2
-				MAX 15
-				HELP "PLLQ value for main PLL.
-
-				It is used to divide VCO output frequency (VCOout) to produce clock for USB OTG FS, SDMMC and RNG
-				(PLLQout). USB OTG FS requires a 48 MHz clock for correct operation, while SDMMC and RNG need a
-				frequency lower than or equal to 48 MHz to work correctly.
-
-				PLLQout = VCOout / PLLQ = PLLin / PLLM * PLLN / PLLQ"
-				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLQ)
-
-		if(CONFIG_CHIP MATCHES "STM32F7[67]")
+		if(distortos_Clocks_07_PLL)
 
 			distortosSetConfiguration(INTEGER
-					distortos_Clocks_10_PLLR
+					distortos_Clocks_08_PLLN
+					336
+					MIN 50
+					MAX 432
+					HELP "PLLN value for main PLL.
+
+					It is used to multiply VCO input frequency (VCOin). Resulting VCO output frequency (VCOout) must be
+					in the range [100 MHz; 432 MHz].
+
+					VCOout = VCOin * PLLN = PLLin / PLLM * PLLN"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLN)
+
+			distortosSetConfiguration(STRING
+					distortos_Clocks_09_PLLP
+					2
+					4
+					6
+					8
+					HELP "PLLP value for main PLL.
+
+					It is used to divide VCO output frequency (VCOout). PLL output frequency (PLLout) must not exceed:
+					- 216 MHz (with over-drive mode) or 180 MHz (otherwise) in voltage scale 1 mode,
+					- 180 MHz (with over-drive mode) or 168 MHz (otherwise) in voltage scale 2 mode,
+					- 144 MHz in voltage scale 3 mode.
+
+					PLLout = VCOout / PLLP = PLLin / PLLM * PLLN / PLLP"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLP
+					OUTPUT_TYPES INTEGER)
+
+			distortosSetConfiguration(INTEGER
+					distortos_Clocks_10_PLLQ
 					7
 					MIN 2
+					MAX 15
+					HELP "PLLQ value for main PLL.
+
+					It is used to divide VCO output frequency (VCOout) to produce clock for USB OTG FS, SDMMC and RNG
+					(PLLQout). USB OTG FS requires a 48 MHz clock for correct operation, while SDMMC and RNG need a
+					frequency lower than or equal to 48 MHz to work correctly.
+
+					PLLQout = VCOout / PLLQ = PLLin / PLLM * PLLN / PLLQ"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLQ)
+
+			if(CONFIG_CHIP MATCHES "STM32F7[67]")
+
+				distortosSetConfiguration(INTEGER
+						distortos_Clocks_11_PLLR
+						7
+						MIN 2
+						MAX 7
+						HELP "PLLR value for main PLL.
+
+						It is used to divide VCO output frequency (VCOout) to produce clock for DSI (PLLRout).
+
+						PLLRout = VCOout / PLLR = PLLin / PLLM * PLLN / PLLR"
+						OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLR)
+
+			endif(CONFIG_CHIP MATCHES "STM32F7[67]")
+
+		endif(distortos_Clocks_07_PLL)
+
+		distortosSetConfiguration(BOOLEAN
+				distortos_Clocks_12_PLLI2S
+				OFF
+				HELP "Enable PLLI2S."
+				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLI2S_ENABLE)
+
+		if(distortos_Clocks_12_PLLI2S)
+
+			distortosSetConfiguration(INTEGER
+					distortos_Clocks_13_PLLI2SN
+					336
+					MIN 50
+					MAX 432
+					HELP "PLLI2SN value for PLLI2S.
+
+					It is used to multiply VCO input frequency (VCOin). Resulting PLLI2S VCO output frequency
+					(PLLI2SVCOout) must be in the range [100 MHz; 432 MHz].
+
+					PLLI2SVCOout = VCOin * PLLI2SN = PLLin / PLLM * PLLI2SN"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLI2SN)
+
+			if(NOT CONFIG_CHIP MATCHES "STM32F7[23]")
+
+				distortosSetConfiguration(STRING
+						distortos_Clocks_14_PLLI2SP
+						2
+						4
+						6
+						8
+						HELP "PLLI2SP value for PLLI2S.
+
+						It is used to divide PLLI2S VCO output frequency (PLLI2SVCOout) to produce clock for SPDIFRX
+						(PLLI2Sout). PLLI2S output frequency must not exceed 216 MHz.
+
+						PLLI2Sout = PLLI2SVCOout / PLLI2SP = PLLin / PLLM * PLLI2SN / PLLI2SP"
+						OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLI2SP
+						OUTPUT_TYPES INTEGER)
+
+			endif(NOT CONFIG_CHIP MATCHES "STM32F7[23]")
+
+			distortosSetConfiguration(INTEGER
+					distortos_Clocks_15_PLLI2SQ
+					2
+					MIN 2
+					MAX 15
+					HELP "PLLI2SQ value for PLLI2S.
+
+					It is used to divide PLLI2S VCO output frequency (PLLI2SVCOout) to produce clock for SAIs
+					(PLLI2SQout). PLLI2S \"Q\" output frequency must not exceed 216 MHz.
+
+					PLLI2SQout = PLLI2SVCOout / PLLI2SQ = PLLin / PLLM * PLLI2SN / PLLI2SQ"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLI2SQ)
+
+			distortosSetConfiguration(INTEGER
+					distortos_Clocks_16_PLLI2SR
+					2
+					MIN 2
 					MAX 7
-					HELP "PLLR value for main PLL.
+					HELP "PLLI2SR value for PLLI2S.
 
-					It is used to divide VCO output frequency (VCOout) to produce clock for DSI (PLLRout).
+					It is used to divide PLLI2S VCO output frequency (PLLI2SVCOout) to produce clock for I2S
+					(PLLI2SRout). PLLI2S \"R\" output frequency must not exceed 192 MHz.
 
-					PLLRout = VCOout / PLLR = PLLin / PLLM * PLLN / PLLR"
-					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLR)
+					PLLI2SRout = PLLI2SVCOout / PLLI2SR = PLLin / PLLM * PLLI2SN / PLLI2SR"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLI2SR)
 
-		endif(CONFIG_CHIP MATCHES "STM32F7[67]")
+		endif(distortos_Clocks_12_PLLI2S)
 
-	endif(distortos_Clocks_04_PLL)
+		distortosSetConfiguration(BOOLEAN
+				distortos_Clocks_17_PLLSAI
+				OFF
+				HELP "Enable PLLSAI."
+				OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLSAI_ENABLE)
+
+		if(distortos_Clocks_17_PLLSAI)
+
+			distortosSetConfiguration(INTEGER
+					distortos_Clocks_18_PLLSAIN
+					288
+					MIN 50
+					MAX 432
+					HELP "PLLSAIN value for PLLSAI.
+
+					It is used to multiply VCO input frequency (VCOin). Resulting PLLSAI VCO output frequency
+					(PLLSAIVCOout) must be in the range [100 MHz; 432 MHz].
+
+					PLLSAIVCOout = VCOin * PLLSAIN = PLLin / PLLM * PLLSAIN"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLSAIN)
+
+			distortosSetConfiguration(STRING
+					distortos_Clocks_19_PLLSAIP
+					2
+					4
+					6 DEFAULT
+					8
+					HELP "PLLSAIP value for PLLSAI.
+
+					It is used to divide PLLSAI VCO output frequency (PLLSAIVCOout) to produce clock for USB, RNG
+					and SDMMC (PLLSAIout). PLLSAI output frequency must not exceed 48 MHz.
+
+					PLLSAIout = PLLSAIVCOout / PLLSAIP = PLLin / PLLM * PLLSAIN / PLLSAIP"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLSAIP
+					OUTPUT_TYPES INTEGER)
+
+			distortosSetConfiguration(INTEGER
+					distortos_Clocks_20_PLLSAIQ
+					2
+					MIN 2
+					MAX 15
+					HELP "PLLSAIQ value for PLLSAI.
+
+					It is used to divide PLLSAI VCO output frequency (PLLSAIVCOout) to produce clock for SAI
+					(PLLSAIQout). PLLSAI \"Q\" output frequency must not exceed 216 MHz.
+
+					PLLSAIQout = PLLSAIVCOout / PLLSAIQ = PLLin / PLLM * PLLSAIN / PLLSAIQ"
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLSAIQ)
+
+			if(NOT CONFIG_CHIP MATCHES "STM32F7[23]")
+
+				distortosSetConfiguration(INTEGER
+						distortos_Clocks_21_PLLSAIR
+						2
+						MIN 2
+						MAX 7
+						HELP "PLLSAIR value for PLLSAI.
+
+						It is used to divide PLLSAI VCO output frequency (PLLSAIVCOout) to produce clock for LCD
+						(PLLSAIRout). PLLSAI \"R\" output frequency must not exceed 216 MHz.
+
+						PLLSAIRout = PLLSAIVCOout / PLLSAIR = PLLin / PLLM * PLLSAIN / PLLSAIR"
+						OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLLSAIR)
+
+			endif(NOT CONFIG_CHIP MATCHES "STM32F7[23]")
+
+		endif(distortos_Clocks_17_PLLSAI)
+
+		if(distortos_Clocks_07_PLL OR distortos_Clocks_17_PLLSAI)
+
+			if(distortos_Clocks_07_PLL)
+				set(pll48ClockSourcePllq PLLQ)
+			endif()
+			if(distortos_Clocks_17_PLLSAI)
+				set(pll48ClockSourcePllsaip PLLSAIP)
+			endif()
+
+			distortosSetConfiguration(STRING
+					distortos_Clocks_22_PLL48_clock_source
+					${pll48ClockSourcePllq}
+					${pll48ClockSourcePllsaip}
+					HELP "Select PLL48 clock source."
+					OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLL48CLK
+					OUTPUT_TYPES BOOLEAN)
+
+		endif(distortos_Clocks_07_PLL OR distortos_Clocks_17_PLLSAI)
+
+	endif(distortos_Clocks_04_PLLs)
 
 	if(distortos_Clocks_03_HSE)
 		set(systemClockSourceHse HSE)
 	endif()
-	if(distortos_Clocks_04_PLL)
+	if(distortos_Clocks_07_PLL)
 		set(systemClockSourcePll PLL DEFAULT)
 	endif()
 
 	distortosSetConfiguration(STRING
-			distortos_Clocks_11_System_clock_source
+			distortos_Clocks_23_System_clock_source
 			HSI
 			${systemClockSourceHse}
 			${systemClockSourcePll}
@@ -182,10 +347,21 @@ else(distortos_Clocks_00_Standard_configuration_of_clocks)
 			- 144 MHz in voltage scale 3 mode."
 			OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_SYSCLK_FREQUENCY)
 
+	distortosSetConfiguration(INTEGER
+			distortos_Clocks_02_PLL48CLK_frequency
+			0
+			MIN 0
+			MAX 48000000
+			HELP "Frequency of PLL48CLK, Hz.
+
+			RCC must be configured by user to achieve that frequency. PLL48CLK frequency must not exceed 48 MHz. Set to
+			0 if PLL48CLK is not available."
+			OUTPUT_NAME CONFIG_CHIP_STM32F7_RCC_PLL48CLK_FREQUENCY)
+
 endif(distortos_Clocks_00_Standard_configuration_of_clocks)
 
 distortosSetConfiguration(STRING
-		distortos_Clocks_12_HPRE
+		distortos_Clocks_24_HPRE
 		1
 		2
 		4
@@ -202,7 +378,7 @@ distortosSetConfiguration(STRING
 		OUTPUT_TYPES INTEGER)
 
 distortosSetConfiguration(STRING
-		distortos_Clocks_13_PPRE1
+		distortos_Clocks_25_PPRE1
 		1
 		2
 		4 DEFAULT
@@ -217,7 +393,7 @@ distortosSetConfiguration(STRING
 		OUTPUT_TYPES INTEGER)
 
 distortosSetConfiguration(STRING
-		distortos_Clocks_14_PPRE2
+		distortos_Clocks_26_PPRE2
 		1
 		2 DEFAULT
 		4